% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{linreg}
\usage{
linreg(formula,data)
}
\description{
Implementing Liear Regression using QR Decomposition
}
\examples{
mod_object <- linreg(Petal.Length~Species, data = iris)
print(mod_object)
resid(mod_object)
}
\references{
\url{http://staff.www.ltu.se/~jove/courses/c0002m/least_squares.pdf}
}
\author{
Hamed , Omid
}
