\name{PackageLab04-package}
\alias{PackageLab04-package}
\alias{PackageLab04}
\docType{package}
\title{
\packageTitle{PackageLab04}
}
\description{
\packageDescription{PackageLab04}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PackageLab04}
\packageIndices{PackageLab04}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{PackageLab04}

Maintainer: \packageMaintainer{PackageLab04}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
